# platform-service/api-docs.yaml
openapi: 3.0.3
info:
  title: Network Measurement Platform - API
  description: |-
    This is the REST API for our application server based on the OpenAPI 3.0 specification. The API is subject to change during the development process.
    
    Useful links:
    - [GitLab Repository](https://gitlab.lrz.de/cm/2023-karthiknarumanchi-mt)
    - [Wiki on GitLab](https://gitlab.lrz.de/cm/2023-karthiknarumanchi-mt/-/wikis/home)
  version: 0.6.0
  contact:
    email: karthik.narumanchi@tum.de
servers:
  - url: localhost:5000
    description: Development server
...
paths:
...
  /api/probes:
    get:
      tags:
      - Probe
      summary: Get all probes with a filter on country and probe status
      operationId: getAllProbes
      parameters:
      - name: country
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - CONNECTED
          - RUNNING
          - DISCONNECTED
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "403":
          description: Forbidden
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProbesResponseObject'
      security:
      - Bearer_Authentication: []
    post:
      tags:
      - Admin
      summary: Create a new probe
      operationId: createProbe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProbeRequestObject'
        required: true
      responses:
        "200":
          description: OK
      security:
      - Bearer_Authentication: []
...
  /api/probes/{probeId}:
    get:
      tags:
      - Probe
      summary: Get a probe by its id
      operationId: getProbeById
      parameters:
      - name: probeId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Probe'
        "403":
          description: Forbidden
        "404":
          description: Probe not found
      security:
      - Bearer_Authentication: []
  /api/probes/my:
    get:
      tags:
      - Probe
      summary: Get all probes for authenticated user with a filter on country and
        probe status
      operationId: getProbesForSelf
      parameters:
      - name: country
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - CONNECTED
          - RUNNING
          - DISCONNECTED
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "403":
          description: Forbidden
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProbesResponseObject'
      security:
      - Bearer_Authentication: []
  /api/probes/config/{probeId}:
    get:
      tags:
      - Probe
      summary: Download the docker-compose config file for a probe
      operationId: getProbeHostFile
      parameters:
      - name: probeId
        in: path
        required: true
        schema:
          type: string
      responses:
        "403":
          description: Forbidden
        "404":
          description: Probe not found
        "200":
          description: OK
          content:
            application/octet-stream: {}
        "401":
          description: Unauthorized
      security:
      - Bearer_Authentication: []
...
components:
  schemas:
    Probe:
      type: object
      properties:
        id:
          type: string
        ipv4:
          type: string
        country:
          type: string
        region:
          type: string
          enum:
          - AF
          - AS
          - EU
          - OC
          - NA
          - SA
        status:
          type: string
          enum:
          - CONNECTED
          - RUNNING
          - DISCONNECTED
        description:
          type: string
        starlinkActiveInterface:
          type: string
        wifiInterfaceActive:
          type: boolean
        wiredInterfaceActive:
          type: boolean
        cellularInterfaceActive:
          type: boolean
...
    ProbeRequestObject:
      required:
      - country
      - description
      - region
      - userEmail
      type: object
      properties:
        userEmail:
          type: string
        country:
          type: string
        region:
          type: string
          enum:
          - AF
          - AS
          - EU
          - OC
          - NA
          - SA
        description:
          type: string
...
    ProbesResponseObject:
      type: object
      properties:
        count:
          type: integer
          format: int64
        probeList:
          type: array
          items:
            $ref: '#/components/schemas/Probe'
        currentPage:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
...
  securitySchemes:
    Bearer_Authentication:
      type: http
      description: A JWT token is required to access this API. JWT token can be obtained
        from the auth/request API.
      scheme: bearer
      bearerFormat: JWT